Nearly have unit tests compiling
