+++ tools/build.sh
@@ -45,6 +45,7 @@ function process_args() {
         ;;
       t)
         UNIT_TESTS=1
+        CMAKE_ARGS+=("-DSPICY_PROTOBUF_TEST=ON")
         ;;
       *)
         usage
@@ -78,24 +79,8 @@ function build {
   #shellcheck disable=SC2048,2086
   "${builder}" ${MAKE_ARGS[*]}
   if [[ "${UNIT_TESTS}" = 1 ]]; then
-    for t in ./tests/*_test; do
-      if [[ -x "${t}" ]]; then
-        echo "============================================="
-        echo "Running ${t}"
-        echo "============================================="
-        "${t}"
-      fi
-    done
-
-    # TODO: Encapsulate tests within 'testing' directory
-    # shellcheck disable=SC2044
-    for binpb in $(find ../test-data/ -name '*.binpb'); do
-      cmd="spicy-driver ./*.hlto -f ${binpb}"
-      echo "============================================="
-      echo "Running ${cmd}"
-      echo "============================================="
-      eval "${cmd}"
-    done
+    pushd "testing/gtest"
+    ctest --progress
     #pushd "../testing/btest" >/dev/null
     #btest -c btest.cfg
     #popd >/dev/null
